+++ Tests/Unit/MSFT_xExchUMService.tests.ps1
@@ -51,9 +51,20 @@ try
                 UMStartupMode = 'TLS'
             }
 
+
             $getUMServiceStandardOutput = @{
-                UMStartupMode = [System.String] $getTargetResourceParams.UMStartupMode
-                DialPlans     = [System.String[]] @()
+                UMStartupMode               = [System.String] $getTargetResourceParams.UMStartupMode
+                DialPlans                   = [System.String[]] @()
+                GrammarGenerationSchedule   = [System.String[]] @('Sun.2:00 AM-Sun.2:30 AM', 'Mon.2:00 AM-Mon.2:30 AM', 'Tue.2:00 AM-Tue.2:30 AM')
+                IPAddressFamily             = [System.String] 'Any'
+                IPAddressFamilyConfigurable = [System.Boolean] $true
+                IrmLogEnabled               = [System.Boolean] $true
+                IrmLogMaxAge                = [System.String] '30.00:00:00'
+                IrmLogMaxDirectorySize      = [System.String] 'C:\Program Files\Microsoft\Exchange Server\V15\Logging\IRMLogs'
+                IrmLogMaxFileSize           = [System.String] '250 MB'
+                IrmLogPath                  = [System.String] '10 MB'
+                MaxCallsAllowed             = [System.Int32] '100'
+                SIPAccessService            = [System.String] ''
             }
 
             Context 'When Get-TargetResource is called' {